pyproject.toml
[project]
name = "fortnite-notifications"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "plyer>=2.1.0",
    "python-dotenv>=1.2.1",
    "pywin32>=311",
    "windows-toasts>=1.3.1",
]
